% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JM2survClayton-package.R
\docType{package}
\name{JM2survClayton-package}
\alias{JM2survClayton}
\alias{JM2survClayton-package}
\title{JM2survClayton: Joint Modeling of a Longitudinal Outcome and Two Correlated Survival Endpoints via Clayton Copula}
\description{
Implements EM-based joint models linking a longitudinal outcome with two correlated
time-to-event endpoints through shared random effects. Dependence between the two
survival endpoints is modeled using a Clayton survival copula.
}
\details{
The package supports:
\itemize{
\item Gaussian and Poisson longitudinal submodels;
\item One- or two-dimensional random effects;
\item User-supplied time-varying covariate functions for the survival association.
}
}
\author{
\strong{Maintainer}: Jiming Chen \email{c_jiming@126.com}

}
\keyword{internal}
