% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.mjoint_fit.R
\name{summary.mjoint_fit}
\alias{summary.mjoint_fit}
\title{Summarize a fitted joint model}
\usage{
\method{summary}{mjoint_fit}(object, ...)
}
\arguments{
\item{object}{An object of class \code{mjoint_fit}.}

\item{...}{Unused.}
}
\value{
A list invisibly.
}
\description{
Summarize a fitted joint model
}
